# docs/openapi.yaml
openapi: 3.1.0
info:
  title: "ZarishSphere FHIR Engine API"
  description: |
    FHIR R5 REST API server for ZarishSphere healthcare platform.
    
    ## Authentication
    All endpoints require a valid SMART on FHIR 2.1 JWT Bearer token.
    See https://docs.zarishsphere.com/auth for details.
    
    ## Versioning
    API version is embedded in the URL: `/fhir/R5/` (FHIR version, not service version).
    
    ## Multi-Tenancy
    All requests are tenant-isolated. The tenant ID is extracted from JWT claims.
  version: "1.0.0"
  contact:
    name: "ZarishSphere Platform"
    email: "platform@zarishsphere.com"
    url: "https://zarishsphere.com"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0"
servers:
  - url: "https://api.zarishsphere.com"
    description: "Production"
  - url: "http://localhost:8001"
    description: "Local development"
security:
  - bearerAuth: []
paths:
...
  /auth/authorize:
    get:
      summary: "OAuth2 authorization endpoint"
      operationId: "authorize"
      tags: ["Authentication"]
      description: "Initiates SMART on FHIR 2.1 authorization flow"
      parameters:
        - name: "response_type"
          in: "query"
          required: true
          schema:
            type: "string"
            enum: ["code"]
        - name: "client_id"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "redirect_uri"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "scope"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "state"
          in: "query"
          schema:
            type: "string"
        - name: "aud"
          in: "query"
          schema:
            type: "string"
      responses:
        "302":
          description: "Redirect to authorization server"
  /auth/token:
    post:
      summary: "OAuth2 token endpoint"
      operationId: "token"
      tags: ["Authentication"]
      description: "Exchanges authorization code for access token"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: "object"
              properties:
                grant_type:
                  type: "string"
                  enum: ["authorization_code", "refresh_token"]
                code:
                  type: "string"
                refresh_token:
                  type: "string"
                client_id:
                  type: "string"
                client_secret:
                  type: "string"
                redirect_uri:
                  type: "string"
      responses:
        "200":
          description: "Token response"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  access_token:
                    type: "string"
                  token_type:
                    type: "string"
                  expires_in:
                    type: "integer"
                  refresh_token:
                    type: "string"
                  scope:
                    type: "string"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
...